REFACTOR: long sync() method to constituent methods.
Syncer.sync() was several dozen lines long but had headers to help
break it up. Some of the components returned values, so I check those
values and bubble them up out of sync. There might be other more intuitive
designs to consider, but for now this is the least change.