Intermediate. Time to go back and rename the Collection **FUNCTION**
Seriously. There is a function. It is named Collection. It is apparently not a class.
I have been trying to derive a class from it for a while now and getting errors that made
no sense in a class declaration. Well, sure.... because the base class I was attempting isn't
on.
one.
Now I am going to go back and rename these.